+++ pkg/datasources/redis/resource/metric_find_value.go
@@ -1,21 +1,19 @@
 package resource
 
 import (
+	"fmt"
 	"regexp"
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
 )
 
-var metricNamesRegex = regexp.MustCompile(`^metrics\(\s*([a-zA-Z0-9._*]*)\s*\)$`)
-var labelValuesRegex = regexp.MustCompile(`^label_values\(\s*([a-zA-Z][a-zA-Z0-9._]*)\s*\)$`)
+var metricNamesRegex = regexp.MustCompile(`^metrics\(\s*([\w.*]*)\s*\)$`)
+var labelNamesRegex = regexp.MustCompile(`^label_names\(\s*([\w.]*)\s*\)$`)
+var labelValuesRegex = regexp.MustCompile(`^label_values\(\s*([\w.]+)\s*\)$`)
 var labelValuesForMetricRegex = regexp.MustCompile(`^label_values\(\s*([a-zA-Z][a-zA-Z0-9._]*)\s*,\s*([a-zA-Z][a-zA-Z0-9._]*)\s*\)$`)
 
 func (rs *Service) getMetricNames(pattern string) ([]MetricFindValue, error) {
-	if pattern == "" {
-		pattern = "*"
-	}
-
-	namesResponse, err := rs.pmseriesAPI.MetricNameMatches(pattern)
+	namesResponse, err := rs.pmseriesAPI.MetricNames(pattern)
 	if err != nil {
 		return nil, err
 	}
@@ -27,12 +25,43 @@ func (rs *Service) getMetricNames(pattern string) ([]MetricFindValue, error) {
 	return metricFindValues, nil
 }
 
+func (rs *Service) getLabelNames(pattern string) ([]MetricFindValue, error) {
+	labelNamesResponse, err := rs.pmseriesAPI.LabelNames(pattern)
+	if err != nil {
+		return nil, err
+	}
+
+	ret := []MetricFindValue{}
+	for _, name := range labelNamesResponse {
+		ret = append(ret, MetricFindValue{name})
+	}
+	return ret, nil
+}
+
+func (rs *Service) getLabelValues(labelName string) ([]MetricFindValue, error) {
+	labelValuesResponse, err := rs.pmseriesAPI.LabelValues([]string{labelName})
+	if err != nil {
+		return nil, err
+	}
+
+	ret := []MetricFindValue{}
+	for _, value := range labelValuesResponse[labelName] {
+		ret = append(ret, MetricFindValue{fmt.Sprintf("%v", value)})
+	}
+	return ret, nil
+}
+
 func (rs *Service) metricFindQuery(query string) ([]MetricFindValue, error) {
 	log.DefaultLogger.Debug("metricFindQuery", "query", query)
 
 	metricNamesQuery := metricNamesRegex.FindStringSubmatch(query)
 	if len(metricNamesQuery) == 2 {
 		return rs.getMetricNames(metricNamesQuery[1])
+	}
+
+	labelNamesQuery := labelNamesRegex.FindStringSubmatch(query)
+	if len(labelNamesQuery) == 2 {
+		return rs.getLabelNames(labelNamesQuery[1])
 	}
 
 	labelValuesQuery := labelValuesRegex.FindStringSubmatch(query)
@@ -42,8 +71,7 @@ func (rs *Service) metricFindQuery(query string) ([]MetricFindValue, error) {
 
 	// deprecated
 	labelValuesForMetricQuery := labelValuesForMetricRegex.FindStringSubmatch(query)
-	log.DefaultLogger.Debug("metricFindQuery", "labelValuesForMetricQuery", labelValuesForMetricQuery)
-
+	log.DefaultLogger.Info("Using deprecated query label_values(metric, label)", "query", query)
 	if len(labelValuesForMetricQuery) == 3 {
 		return rs.getLabelValues(labelValuesForMetricQuery[2])
 	}